refactor(register): refactor handleRegister event
